Moved TextDecoder into the Components/ subdir
